+++ R/RunMAPOneSample.R
@@ -99,7 +99,7 @@ RunMAPOnOneSample <-
     
     
     if (!is.null(mut.type)) {
-      ets <- PCAWG7::GetEtiology(mut.type, colnames(sigs1))
+      ets <- cosmicsig::get_etiology(mut.type, colnames(sigs1))
       colnames(sigs1) <-
         paste0(colnames(sigs1), " (exposure = ", round(inferred.exposure[, 1]),
                ", contribution = ",