+++ R/ShowSigActivity.R
@@ -24,7 +24,7 @@
 #' \dontrun{
 #' spectra <- PCAWG7::spectra$PCAWG$SBS96[, 1:2, drop = FALSE]
 #' exposure <- PCAWG7::exposure$PCAWG$SBS96[, 1:2, drop = FALSE]
-#' sigs <- PCAWG7::signature$genome$SBS96
+#' sigs <- cosmicsig::COSMIC_v3.2$signature$GRCh37
 #' sigs.prop <- ExposureProportions(mutation.type = "SBS96", 
 #'                                  cancer.type = "Biliary-AdenoCA")
 #' retval <- AddSigActivity(spectra, exposure, sigs, sigs.prop)