+++ tests/testthat/test_ShowSigActivity.R
@@ -5,7 +5,7 @@ test_that("ShowSigActivity for only one SBS96 catalog", {
   spect.name <- colnames(spect)
   index <- grep(spect.name, x = colnames(PCAWG7::exposure$PCAWG$SBS96))
   exposure <- PCAWG7::exposure$PCAWG$SBS96[, index, drop = FALSE]
-  sigs <- PCAWG7::signature$genome$SBS96
+  sigs <- cosmicsig::COSMIC_v3.2$signature$GRCh37$SBS96
   sigs.prop <- ExposureProportions(mutation.type = "SBS96",
                                    cancer.type = "Biliary-AdenoCA")
   retval <- AddSigActivity(spect = spect, exposure = exposure, sigs = sigs,
@@ -31,7 +31,7 @@ test_that("ShowSigActivity for multiple SBS96 catalog", {
   indices <-
     sapply(spect.name, FUN = grep, x = colnames(PCAWG7::exposure$PCAWG$SBS96))
   exposure <- PCAWG7::exposure$PCAWG$SBS96[, indices, drop = FALSE]
-  sigs <- PCAWG7::signature$genome$SBS96
+  sigs <- cosmicsig::COSMIC_v3.2$signature$GRCh37$SBS96
   sigs.prop <- ExposureProportions(mutation.type = "SBS96",
                                    cancer.type = "Biliary-AdenoCA")
   retval <- AddSigActivity(spect = spect, exposure = exposure, sigs = sigs,
@@ -55,7 +55,7 @@ test_that("ShowSigActivity for only one DBS78 catalog", {
   spect.name <- colnames(spect)
   index <- grep(spect.name, x = colnames(PCAWG7::exposure$PCAWG$DBS78))
   exposure <- PCAWG7::exposure$PCAWG$DBS78[, index, drop = FALSE]
-  sigs <- PCAWG7::signature$genome$DBS78
+  sigs <- cosmicsig::COSMIC_v3.2$signature$GRCh37$DBS78
   sigs.prop <- ExposureProportions(mutation.type = "DBS78",
                                    cancer.type = "Biliary-AdenoCA")
   retval <- AddSigActivity(spect = spect, exposure = exposure, sigs = sigs,
@@ -80,7 +80,7 @@ test_that("ShowSigActivity for multiple DBS78 catalog", {
   indices <-
     sapply(spect.name, FUN = grep, x = colnames(PCAWG7::exposure$PCAWG$DBS78))
   exposure <- PCAWG7::exposure$PCAWG$DBS78[, indices, drop = FALSE]
-  sigs <- PCAWG7::signature$genome$DBS78
+  sigs <- cosmicsig::COSMIC_v3.2$signature$GRCh37$DBS78
   sigs.prop <- ExposureProportions(mutation.type = "DBS78",
                                    cancer.type = "Biliary-AdenoCA")
   retval <- AddSigActivity(spect = spect, exposure = exposure, sigs = sigs,
@@ -104,7 +104,7 @@ test_that("ShowSigActivity for only one ID catalog", {
   spect.name <- colnames(spect)
   index <- grep(spect.name, x = colnames(PCAWG7::exposure$PCAWG$ID))
   exposure <- PCAWG7::exposure$PCAWG$ID[, index, drop = FALSE]
-  sigs <- PCAWG7::signature$genome$ID
+  sigs <- cosmicsig::COSMIC_v3.2$signature$GRCh37$ID
   sigs.prop <- ExposureProportions(mutation.type = "ID",
                                    cancer.type = "Biliary-AdenoCA")
   retval <- AddSigActivity(spect = spect, exposure = exposure, sigs = sigs,
@@ -129,7 +129,7 @@ test_that("ShowSigActivity for multiple ID catalog", {
   indices <-
     sapply(spect.name, FUN = grep, x = colnames(PCAWG7::exposure$PCAWG$ID))
   exposure <- PCAWG7::exposure$PCAWG$ID[, indices, drop = FALSE]
-  sigs <- PCAWG7::signature$genome$ID
+  sigs <- cosmicsig::COSMIC_v3.2$signature$GRCh37$ID
   sigs.prop <- ExposureProportions(mutation.type = "ID",
                                    cancer.type = "Biliary-AdenoCA")
   retval <- AddSigActivity(spect = spect, exposure = exposure, sigs = sigs,